docs(charts,symbology): add resolve_categorical_cmap examples; fix bar_by doctest repr
- document resolve_categorical_cmap with executable examples (the qualitative-default swap)
- pin the bar_by doctest to plain floats so it is stable under numpy 2's np.float64 repr

+++ src/digitalearth/charts.py
@@ -437,7 +437,7 @@ def bar_by(data: Any, by: str, column: Optional[str] = None, *, agg: str = "sum"
             ...     crs=4326,
             ... )
             >>> ax = bar_by(gdf, "cat", "v", agg="sum")
-            >>> [round(rect.get_height(), 1) for rect in ax.containers[0]]
+            >>> [float(round(rect.get_height(), 1)) for rect in ax.containers[0]]
             [3.0, 3.0]
 
             ```
